Fix pow() func using int as accumulator.
CesiumMan.gltf exposed the pow func issue, plus the e issue
in the previous commit. This commit fixes the pow issue
and updates gtlf accessor tests and test_gltf.gltf file
with catches for the pow func and the e issue.

+++ math.hpp
@@ -9,9 +9,9 @@ static inline float rad(float deg) {
     return 0.0055555555 * 3.1415927 * deg;
 }
 
-static inline int pow(int num, int exp) {
-    int accum = 1;
-    for(int i = 0; i < exp; ++i) {
+static inline u64 pow(u64 num, u32 exp) {
+    u64 accum = 1;
+    for(u32 i = 0; i < exp; ++i) {
         accum *= num;
     }
     return accum;
